refactor: Organize into packages
Code was messy.

Divide classes into packages.

Let's create duke duke.task, duke.command, duke.exception packages.
