+++ .gitignore
@@ -16,5 +16,5 @@ bin/
 /text-ui-test/ACTUAL.TXT
 text-ui-test/EXPECTED-UNIX.TXT
 
-# Duke Tasks File
+# duke.Duke Tasks File
 /data/
