+++ src/main/java/duke/exception/InvalidArgumentException.java
@@ -1,3 +1,5 @@
+package duke.exception;
+
 public class InvalidArgumentException extends DukeException {
     private String command;
