+++ src/main/java/duke/exception/NoTaskFoundException.java
@@ -1,3 +1,5 @@
+package duke.exception;
+
 public class NoTaskFoundException extends DukeException {
     private String taskId;
