+++ src/main/java/duke/exception/TasksFileException.java
@@ -1,3 +1,5 @@
+package duke.exception;
+
 public class TasksFileException extends DukeException {
     @Override
     public String toString() {
